Add implementation of voters signature based on OpenSSL
The core feature of "Sam sabie CVK" project is voter's signature.
It makes possible for an alternative election board to be anonymous,
fair and transparent.

The voter's signature has three distinctive properties:
    1. it is impossible to know the legal identity from the voter's signature;
    2. it is impossible to find a voter's signature from legal identity;
    3. two voter's signatures can be compared to having the same legal
       identities without decrypting them.

This patch makes draft implementation of the voter's signature based on the
socialist millionaire problem. The implementation uses modular arithmetic and
properties of exponents in particular. The libcrypto library from OpenSSL is
used for long and modular arithmetic and cryptosecure random numbers. The code
follows OpenSSL design patterns.

The patch contains three major functions: preparation of public keys, creating
a voter's signature and test for the same legal identity for two voter's
signatures. However, we expect that all pairs comparison requires a significant
performance boost to become feasible for a realistic number of users, hence some
algorithmic optimizations are possibly needed for the case.

The patch adds also small programs as example of keys and voter's signatures
and theirs comparison. It can be built via `make`. The current Makefile is not
general enough as should be and will be fixed in the future. The gitignore file
was updated to match new build structure.