Add functionality to obtain mapped identity
The pairwise comparison is not feasible on expected numbers of voters
(there is O(n^2) comparison). It was found that some values from P and Q
can be obtained that uniquely binded with the legal identity, what can
be used for fast check of duplication. However, it is required to keep
`k2` key secreet, to allow EB make a comparison but prohibit the third
party map from legal identity to voter's signature.

This patch add a new function to calculate a maped legal identity used
inside EB. The documentation for other functons was changed approprietaly
to reflect change in the protocol understanding. Also a new strucutre was
added to keep voter's signature and mapped legal identity together.