Add treap to verify hyothesis about performance
There is no intention to use this code in production, but it allows to keep
crypto related part C only (as OpenSSL is pure C). In production some reliable
library implementation of hash map will be used. The code here allows to prove
with programm change in performance when mapped identity is used and comparison
is done using fast access structure like hash maps or binary search tree.
The treap has only on average assymptotic O(log N)  for insertions and other
operations, however, it is much simpler than RB-tree and hash map. So it seems
appropriate to use treap for test purposes. Note that this is ternary treap, thus
duplicates of ballots are marked by insertion automatically.